endpoint list-books

+++ books/models.py
@@ -5,7 +5,7 @@ from django.db import models
 class book(models.Model):
     title = models.CharField(max_length=100)
     author = models.CharField(max_length=100)
-    price = models.FloatField()
+    price = models.DecimalField(max_digits=5, decimal_places=2)
     created_at = models.DateTimeField(auto_now_add=True)
     updated_at = models.DateTimeField(auto_now=True)
     available = models.BooleanField(default=True)
